namespace renamings
